write doc reconstructure sync.rs split sync process to collection and media sync
